added listing factory and seeder

+++ database/migrations/2025_08_26_113119_add_fields_to_listings_table.php
@@ -13,14 +13,14 @@ return new class extends Migration {
         Schema::table('listings', function (Blueprint $table) {
             $table->unsignedTinyInteger('beds');
             $table->unsignedTinyInteger('baths');
-            $table->unsignedTinyInteger('area');
+            $table->unsignedInteger('area');
 
             $table->tinyText('city');
             $table->tinyText('code');
             $table->tinyText('street');
             $table->tinyText('street_nr');
 
-            $table->unsignedTinyInteger('price');
+            $table->unsignedInteger('price');
         });
     }
